Adding support for Spark322
This patch adds Spark322 support
