+++ backends-velox/src/main/scala/io/glutenproject/execution/VeloxHashAggregateExecTransformer.scala
@@ -371,6 +371,8 @@ case class VeloxHashAggregateExecTransformer(
     context.registerAggregationParam(operatorId, aggParams)
     resRel
   }
+  override def isStreaming: Boolean = false
+  def numShufflePartitions: Option[Int] = Some(0)
   override protected def withNewChildInternal(newChild: SparkPlan)
     : VeloxHashAggregateExecTransformer = {
       copy(child = newChild)
