+++ pom.xml
@@ -37,7 +37,7 @@
   <properties>
     <scala.version>2.12.10</scala.version>
     <scala.binary.version>2.12</scala.binary.version>
-    <spark.version>3.2.1</spark.version>
+    <spark.version>3.2.2</spark.version>
     <arrow.version>8.0.0-gluten-SNAPSHOT</arrow.version>
     <arrow-memory.artifact>arrow-memory-unsafe</arrow-memory.artifact>
     <hadoop.version>${hadoop.version}</hadoop.version>
@@ -64,7 +64,7 @@
     <project.prefix>spark-sql-columnar</project.prefix>
     <project.name.prefix>Gluten</project.name.prefix>
     <jar.assembly.name.prefix>gluten</jar.assembly.name.prefix>
-    <spark321.version>3.2.1</spark321.version>
+    <spark322.version>3.2.2</spark322.version>
   </properties>
 
   <profiles>
@@ -74,13 +74,13 @@
           <activeByDefault>true</activeByDefault>
       </activation>
       <properties>
-          <spark.version>${spark321.version}</spark.version>
+          <spark.version>${spark322.version}</spark.version>
       </properties>
     </profile>
     <profile>
-        <id>spark-3.2.1</id>
+        <id>spark-3.2.2</id>
         <properties>
-            <spark.version>${spark321.version}</spark.version>
+            <spark.version>${spark322.version}</spark.version>
         </properties>
     </profile>
     <profile>
@@ -109,7 +109,7 @@
     <profile>
       <id>emr-6.3.0</id>
       <properties>
-        <hadoop.version>3.2.1</hadoop.version>
+        <hadoop.version>3.2.2</hadoop.version>
       </properties>
     </profile>
     <profile>
