+++ shims/common/pom.xml
@@ -77,7 +77,7 @@
       <dependency>
           <groupId>org.apache.spark</groupId>
           <artifactId>spark-sql_${scala.binary.version}</artifactId>
-          <version>${spark321.version}</version>
+          <version>${spark322.version}</version>
     <scope>provided</scope>
       </dependency>
   </dependencies>
