--- shims/spark321/src/main/resources/META-INF/services/io.glutenproject.sql.shims.SparkShimProvider
@@ -1,1 +1,0 @@
-io.glutenproject.sql.shims.spark321.SparkShimProvider
