+++ shims/spark322/src/main/resources/META-INF/services/io.glutenproject.sql.shims.SparkShimProvider
@@ -1,0 +1,1 @@
+io.glutenproject.sql.shims.spark322.SparkShimProvider
