+++ shims/spark322/src/main/scala/io/glutenproject/sql/shims/spark322/Spark322Shims.scala
@@ -14,21 +14,10 @@
  * limitations under the License.
  */
 
-package io.glutenproject.sql.shims.spark321
+package io.glutenproject.sql.shims.spark322
 
-import io.glutenproject.sql.shims.{SparkShims, SparkShimDescriptor}
+import io.glutenproject.sql.shims.{SparkShims, ShimDescriptor}
 
-object SparkShimProvider {
-  val DESCRIPTOR = SparkShimDescriptor(3, 2, 1)
-  val DESCRIPTOR_STRINGS = Seq(s"$DESCRIPTOR")
-}
-
-class SparkShimProvider extends io.glutenproject.sql.shims.SparkShimProvider {
-  def createShim: SparkShims = {
-    new Spark321Shims()
-  }
-
-  def matches(version: String): Boolean = {
-    SparkShimProvider.DESCRIPTOR_STRINGS.contains(version)
-  }  
+class Spark322Shims extends SparkShims {
+  override def getShimDescriptor: ShimDescriptor = SparkShimProvider.DESCRIPTOR
 }
