+++ shims/spark322/src/test/scala/io/glutenproject/sql/shims/spark322/Spark322ShimsSuite.scala
@@ -14,10 +14,21 @@
  * limitations under the License.
  */
 
-package io.glutenproject.sql.shims.spark321
+package io.glutenproject.sql.shims.spark322
 
-import io.glutenproject.sql.shims.{SparkShims, ShimDescriptor}
+import io.glutenproject.sql.shims.{SparkShims, SparkShimLoader, SparkShimDescriptor}
 
-class Spark321Shims extends SparkShims {
-  override def getShimDescriptor: ShimDescriptor = SparkShimProvider.DESCRIPTOR
+import org.scalatest.FunSuite;
+
+class Spark321ShimsSuite extends FunSuite {
+  val descriptor = SparkShimDescriptor(3, 2, 2)
+  test("Spark shims descriptor") {
+    val sparkShims: SparkShims = new SparkShimProvider().createShim
+    assert(sparkShims.getShimDescriptor === descriptor)
+  }
+  
+  test("Spark shims loader") {
+    val sparkShims: SparkShims = SparkShimLoader.getSparkShims
+    assert(sparkShims.getShimDescriptor === descriptor)
+  }
 }
